test-python/langchain/requirements.txt
jupyter>=1.0.0
jupyterlab
langchain[docarray]==0.0.284
langchain_experimental==0.0.25
langsmith==0.0.65
huggingface_hub>=0.14.1
openai==0.28.1
google-cloud-aiplatform>=1.26.0
google-generativeai
replicate
ruff>=0.0.278
pypdf>=3.12.2
tiktoken>=0.4.0
duckduckgo_search>=3.8.4
streamlit>=1.24.1
arxiv>=1.4.8
wikipedia>=1.4.0
wolframalpha>=5.0.0
unstructured>=0.8.1
pandoc>=1.1.0  # needs system install as well!
sentence-transformers>=2.2.2
gradio>=3.38.0
gpt4all>=1.0.6
google-api-python-client>=2.94.0
apify-client>=1.4.0
chromadb>=0.4.8
lanarky==0.7.16
ray[serve]>=2.6.3
faiss-cpu>=1.7.4
promptwatch>=0.2.9
langchain_decorators>=0.2.3